Create Create Perspective From Fact Tables.csx
This script will take the selected tables and create a perspective that includes all the related tables, measures, & calc groups. It will take into consideration items that are nested. This script is useful for adding context to large models with multiple fact tables, letting you narrow down the view for an end user, so they are not overwhelmed.